requirement.txt
flask_cors
requests
python-dotenv
inquirer
